+++ modules/ghlint-testing/src/main/kotlin/net/twisterrob/ghlint/testing/matcher.kt
@@ -7,35 +7,23 @@ import net.twisterrob.ghlint.results.Finding
 public fun beEmpty(): Matcher<List<Finding>> = object : Matcher<List<Finding>> {
 	override fun test(value: List<Finding>): MatcherResult = MatcherResult(
 		value.isEmpty(),
-		{
-			val values = value.joinToString(separator = "\n", transform = Finding::testString)
-			"Findings should be empty but contained:\n$values"
-		},
+		{ "Findings should be empty but contained:\n${value.testString()}" },
 		{ "Collection should not be empty" }
 	)
 }
 
-public fun Finding.testString(): String {
-	val coordinates = with(location) {
-		"${start.line.number}:${start.column.number}-${end.line.number}:${end.column.number}"
-	}
-	return """
-		|Finding(
-		|	rule=${rule},
-		|	issue=${issue.id},
-		|	location=${location.file.path}/${coordinates},
-		|	message=${message}
-		|)
-	""".trimMargin()
+public fun haveFinding(issue: String): Matcher<List<Finding>> = object : Matcher<List<Finding>> {
+	override fun test(value: List<Finding>): MatcherResult = MatcherResult(
+		value.singleOrNull { it.issue.id == issue } != null,
+		{ "Could not find ${issue} among findings:\n${value.testString()}" },
+		{ "Collection should not have ${issue}." }
+	)
 }
 
 public fun haveFinding(issue: String, message: String): Matcher<List<Finding>> = object : Matcher<List<Finding>> {
 	override fun test(value: List<Finding>): MatcherResult = MatcherResult(
 		value.singleOrNull { it.issue.id == issue && it.message == message } != null,
-		{
-			val values = value.joinToString(separator = "\n", transform = Finding::testString)
-			"Could not find ${issue}: ${message} among findings:\n${values}"
-		},
+		{ "Could not find ${issue}: ${message} among findings:\n${value.testString()}" },
 		{ "Collection should not have ${issue}: ${message}." }
 	)
 }
